Separate reference calculation from absorber implementation
